+++ benchmark/test_unary_pointwise_perf.py
@@ -3,7 +3,13 @@ from typing import Generator
 import pytest
 import torch
 
-from .attri_util import BOOL_DTYPES, DEFAULT_METRICS, FLOAT_DTYPES, INT_DTYPES
+from .attri_util import (
+    BOOL_DTYPES,
+    COMPLEX_DTYPES,
+    DEFAULT_METRICS,
+    FLOAT_DTYPES,
+    INT_DTYPES,
+)
 from .performance_utils import Benchmark, generate_tensor_input, vendor_name
 
 
@@ -31,6 +37,7 @@ class UnaryPointwiseBenchmark(Benchmark):
 
 forward_operations = [
     ("abs", torch.abs, FLOAT_DTYPES),
+    ("angle", torch.angle, COMPLEX_DTYPES + [torch.float32]),
     ("erf", torch.erf, FLOAT_DTYPES),
     ("exp", torch.exp, FLOAT_DTYPES),
     ("neg", torch.neg, FLOAT_DTYPES),
