+++ tests/test_unary_pointwise_ops.py
@@ -7,6 +7,7 @@ from .accuracy_utils import (
     ALL_FLOAT_DTYPES,
     ALL_INT_DTYPES,
     BOOL_TYPES,
+    COMPLEX_DTYPES,
     FLOAT_DTYPES,
     INT_DTYPES,
     POINTWISE_SHAPES,
@@ -45,6 +46,18 @@ def test_accuracy_abs_(shape, dtype):
     with flag_gems.use_gems():
         res_out = torch.abs_(inp)
 
+    gems_assert_equal(res_out, ref_out)
+
+
+@pytest.mark.angle
+@pytest.mark.parametrize("shape", POINTWISE_SHAPES)
+@pytest.mark.parametrize("dtype", COMPLEX_DTYPES + [torch.float32])
+def test_accuracy_angle(shape, dtype):
+    inp = torch.randn(shape, dtype=dtype, device=flag_gems.device)
+
+    ref_out = torch.angle(inp)
+    with flag_gems.use_gems():
+        res_out = torch.angle(inp)
     gems_assert_equal(res_out, ref_out)
